Extract popup comment composer from content script
